backend/requirements.txt
flask==3.0.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
PyMuPDF==1.24.0
chromadb==0.5.3
openai>=1.0.0
httpx>=0.25.0
python-dotenv==1.0.0
werkzeug==3.0.1
